+++ src/Services/LicenseService/AppyNox.Services.License.WebAPI/Program.cs
@@ -18,6 +18,8 @@ using Microsoft.OpenApi.Models;
 using Serilog;
 using AppyNox.Services.Base.API.Authentication;
 using AppyNox.Services.Base.Application.Interfaces.Authentication;
+using AppyNox.Services.Authentication.WebAPI.Permission;
+using Microsoft.AspNetCore.Authentication;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -84,8 +86,8 @@ builder.Services.AddSingleton(jwtConfiguration);
 
 builder.Services.AddAuthentication(options =>
 {
-    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+    options.DefaultAuthenticateScheme = "NoxJwtScheme";
+    options.DefaultChallengeScheme = "NoxJwtScheme";
 }).AddJwtBearer(options =>
 {
     options.TokenValidationParameters = new TokenValidationParameters
@@ -98,6 +100,9 @@ builder.Services.AddAuthentication(options =>
         ValidAudience = jwtConfiguration.Audience,
         IssuerSigningKey = new SymmetricSecurityKey(jwtConfiguration.GetSecretKeyBytes())
     };
+})
+.AddScheme<AuthenticationSchemeOptions, NoxJwtAuthenticationHandler>("NoxJwtScheme", options =>
+{
 });
 
 if (!builder.Environment.IsDevelopment())
@@ -148,7 +153,7 @@ builder.Services.AddAuthorization(options =>
     }
 });
 
-builder.Services.AddScoped<IAuthorizationHandler, PermissionAuthorizationHandler>();
+builder.Services.AddScoped<IAuthorizationHandler, NoxJwtAuthorizationHandler>();
 builder.Services.AddScoped<INoxTokenManager, NoxTokenManager>();
 noxLogger.LogInformation("Registering JWT Configuration completed.");
 
@@ -164,6 +169,9 @@ if (!app.Environment.IsDevelopment())
     app.UseSwaggerUI();
 }
 
+app.UseApiResponseAndExceptionWrapper(new AutoWrapperOptions { IsApiOnly = true, ShowApiVersion = true, ApiVersion = "1.0", ShowStatusCode = true });
+app.UseMiddleware<ExceptionHandlingMiddleware>();
+
 app.UseMiddleware<CorrelationIdMiddleware>();
 
 app.UseHttpsRedirection();
@@ -176,8 +184,6 @@ app.UseMiddleware<UserIdMiddleware>();
 
 app.MapControllers();
 
-app.UseApiResponseAndExceptionWrapper(new AutoWrapperOptions { IsApiOnly = true, ShowApiVersion = true, ApiVersion = "1.0", ShowStatusCode = true });
-app.UseMiddleware<ExceptionHandlingMiddleware>();
 app.UseMiddleware<QueryParameterValidateMiddleware>();
 
 app.UseHealthChecks("/api/health");